+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [unreleased]
+* Support for short and tick barlines  with `measure@bar.len` and `measure@bar.place` (@earboxer)
+* Support for dashed and dotted slurs and ties ((@earboxer and @napulen)
 * Options for controlling system divider display (--systemDivider "none|left|left-right")
 
 ## [2.3.3] - 2019-11-26