+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [unreleased]
+* Fig coloring of ties and slurs (@rettinghaus)
+* Fig bug with tuplet number placement (@rettinghaus)
 * Support for short and tick barlines  with `measure@bar.len` and `measure@bar.place` (@earboxer)
 * Support for dashed and dotted slurs and ties ((@earboxer and @napulen)
 * Options for controlling system divider display (--systemDivider "none|left|left-right")