+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Options for controlling system divider display (--systemDivider "none|left|left-right")
 * Fix coloring of ties and slurs (@rettinghaus)
 * Fix bug with tuplet number placement (@rettinghaus)
+* LibMEI output generated from MEI 4.0.1
 
 ## [2.3.3] - 2019-11-26
 * Fix bug with NPM build (remove init function for adjusting memory)